The Key to Hire an Effective Virtual Assistance
Hiring a Virtual  Assistant plays an important role on the success of a business. They are the one doing the less important role of a business owner so that they can focus on more important task, such as business development. However, hiring an efficient VA may be a challenge to some start up entrepreneurs. Here are some tips to help you find an efficient one:


Post Your Job Opening to a Best Site

It is recommended to have you post your job opening so that you can hire the best employee that you need. There are several outsourcing platforms that can serve you well in providing Virtual Assistant candidates. I personally recommend Upwork.com and OnlineJobs.ph, you can find a lot of candidates with specified skills and experience that can definitely help you choose on a project basis, full-time or part-time. 


Identify the Job Description

In hiring an employee you must consider the details of the work that you will assign, it is the same thing you will apply in hiring a virtual assistant. 
     1. The number of hours that he/she will work to accomplish your assigned task efficiently

     2. The task should be done in a specific time you allotted or in a daily basis

     3. The skills and experience a VA must have to provide the services you are expecting

     4. The equipment he/she must have in order to accomplish your assigned task

     5. What specific work values you are expecting them to have

     6. The amount of task you will initially give and then later on you will increase based on the performance  



Specify Your Budget

You must be mindful on the budget of hiring a virtual assistant. This is also not to waste your time on outsourcing a candidate that you can't afford the rate. A discussion is necessary to meet the goal effectively, especially if you are hiring on a project basis. If you need an assistant for full-time or part-time basis, if necessary allot a budget to give an incentives based on the performance for a job well done. 



Conduct an Interview

You must know what skills and qualities a prospective candidate must possess for you to work compatibly. Look for a virtual assistant who knows and understands your business, make sure they understand your target market, and make sure they understand social media. The interview process must not focus on their prior work experience but on how they will handle a specific situations, and most importantly is that they are willing to learn and accept feedback. 





The process in selecting a Virtual Assistant is very crucial as they will be a part of your team in the crucial stage of the development of your business. It is beneficial also for you to be aware on the activity of the shortlisted candidate based on what they posted on their social media accounts. Most of the time you can determine their previous projects and clients tagged on their profile, the people associating them and their work experiences, and the services they are offering to public. It is not necessary, but the serious applicants are normally cautious on the impression they will build to public as what is being shown on their Facebook account/page, instagram, twitter, linkedIn profile. If they can maintain their positive impression to the public, most probably they will also handle your company's public impression cautiously. 



Want to try a Virtual Assistant fit to your needs?

You may want to consider checking this site for the services and portfolios that will help you decide on hiring an efficient and effective Virtual Assistant. 

https://eabusinessconsultant.wixsite.com/ernaargullano